fix(Condition) : fix fields data type long to int ErrorType과 Procedure에서 Integer 리스트를 받기 때문에 더이상 long이 아닌 int로 수정함.

+++ src/main/java/lotto/constant/Condition.java
@@ -1,13 +1,13 @@
 package lotto.constant;
 
 public final class Condition {
-    public static final long MIN_LOTTO_MONEY = 0;
-    public static final long THOUSAND = 1_000;
+    public static final int MIN_LOTTO_MONEY = 0;
+    public static final int THOUSAND = 1_000;
 
-    public static final long SIX_WINNING_NUMBERS = 6;
-    public static final long ONE_NUMBER = 1;
+    public static final int SIX_WINNING_NUMBERS = 6;
+    public static final int ONE_NUMBER = 1;
 
-    public static final long MIN_DRAW_NUMBER = 1;
-    public static final long MAX_DRAW_NUMBER = 45;
+    public static final int MIN_DRAW_NUMBER = 1;
+    public static final int MAX_DRAW_NUMBER = 45;
 
 }
